Modification token et ajouts validation images back

+++ trouve_ta_voie_serveur/controllers/imageController.js
@@ -2,22 +2,37 @@
 
 const db = require("../models/dbSetup");
 const Image = db.images;
+const Grimpe = db.grimpes;
 
 const uuid = require("uuid");
 const fs = require("fs");
 
 exports.deleteImgById = async (req, res) => {
-    // todo : Verifie que le user connecter peut supprimer cette image
     try {
         let resultImg = await Image.findByPk(req.params.id);
 
         if (resultImg) {
-            await Image.destroy({where: {id: req.params.id}}).then(() => {
-                fs.unlinkSync(`public/${resultImg.dataValues.path}`);
-                res.status(204).end();
+            let grimpe = await Grimpe.findOne({
+                where: {id: resultImg.grimpeId},
+                include: [Image],
             });
+
+            // Dois avoir au moins une image.
+            if (grimpe && grimpe.images.length <= 1) {
+                res.status(400).end(); // edit: Verifie le code a retourne
+            }
+
+            // Est autorisé.
+            if (grimpe.utilisateurId === req.token.userId || req.token.isAdmin) {
+                await Image.destroy({where: {id: req.params.id}}).then(() => {
+                    fs.unlinkSync(`public/${resultImg.dataValues.path}`);
+                    res.status(204).end();
+                });
+            } else {
+                res.status(403).end();
+            }
         } else {
-            res.stat(404).end();
+            res.status(404).end();
         }
     } catch (e) {
         res.status(500).end();
@@ -25,28 +40,41 @@ exports.deleteImgById = async (req, res) => {
 };
 
 exports.addImg = async (req, res) => {
-    // todo : Check si le user peut ajoute l<image a la grimpe
     try {
         await db.sequelize.transaction(async (transaction) => {
             try {
-                const imgb64 = req.body.imgBase64;
-                const matches = imgb64.base64.match(/^data:([A-Za-z+/]+);base64,(.+)$/);
-                let buff = Buffer.from(matches[2], "base64");
-                const dotIndex = imgb64.name.lastIndexOf('.');
-                const fileName = uuid.v4() + imgb64.name.substring(dotIndex);
+                let grimpe = await Grimpe.findOne({
+                    where: {id: req.body.grimpeId},
+                });
 
-                fs.writeFileSync(`public/img/grimpe/${fileName}`, buff);
+                // Grimpe n'existe pas.
+                if (!grimpe) {
+                    res.status(404).end();
+                }
 
-                let img = await Image.create({
-                    nom: fileName,
-                    path: `/img/grimpe/${fileName}`,
-                    grimpeId: req.body.grimpeId,
-                }, {transaction: transaction});
+                // Est autorisé.
+                if (grimpe.utilisateurId === req.token.userId || req.token.isAdmin) {
+                    const imgb64 = req.body.imgBase64;
+                    const matches = imgb64.base64.match(/^data:([A-Za-z+/]+);base64,(.+)$/);
+                    let buff = Buffer.from(matches[2], "base64");
+                    const dotIndex = imgb64.name.lastIndexOf(".");
+                    const fileName = uuid.v4() + imgb64.name.substring(dotIndex);
 
-                if (img) {
-                    res.status(201).json(img);
+                    fs.writeFileSync(`public/img/grimpe/${fileName}`, buff);
+
+                    let img = await Image.create({
+                        nom: fileName,
+                        path: `/img/grimpe/${fileName}`,
+                        grimpeId: req.body.grimpeId,
+                    }, {transaction: transaction});
+
+                    if (img) {
+                        res.status(201).json(img);
+                    } else {
+                        res.status(400).end();
+                    }
                 } else {
-                    res.status(400).end();
+                    res.status(403).end();
                 }
             } catch (e) {
                 res.status(400).end();
